gvconfig: squash a -Wsign-conversion warning

+++ lib/gvc/gvconfig.c
@@ -605,7 +605,7 @@ void gvconfig(GVC_t * gvc, boolean rescan)
 		return;
     	    }
     	    else {
-    	        config_text = gmalloc(config_st.st_size + 1);
+    	        config_text = gmalloc((size_t)config_st.st_size + 1);
     	        sz = fread(config_text, 1, config_st.st_size, f);
     	        if (sz == 0) {
     		    agerr(AGERR,"%s is zero sized, or other read error.\n", gvc->config_path);
